Externalise provider request method
